:bug: fix(v1.0.0 amend): fix bug that get hex string from bytes

+++ src/main/java/weaver/micro/devkit/codec/CodecUtils.java
@@ -9,7 +9,8 @@ public class CodecUtils {
     public static String binary2Hex(byte[] bytes) {
         StringBuilder result = new StringBuilder(bytes.length << 1);
 
-        for (byte aByte : bytes) {
+        // 此处不能用byte类型接收, 除非是unsigned, 但显然不可能, 所以用short代替
+        for (short aByte : bytes) {
             if (aByte < 0) aByte += 256;
 
             if (aByte < 16) result.append("0");
